+++ lib/ex_mcp/session_manager.ex
@@ -1221,7 +1221,7 @@ defmodule ExMCP.SessionManager do
   end
 
   defp open_store!(config) do
-    case SessionStore.open(config) do
+    case SessionStore.Factory.open(config) do
       {:ok, store} ->
         store